remove old router and rename review router
